Move EmptyMetricData to internal package
